refactor(fuzz): extract shared streaming utilities into reusable module
- Move split_sizes_from_seed, convert, and normalize_ascii_whitespace
  into streaming_utils.rs shared by both chunk-split fuzz targets
- Add content-type toggle (first-byte flag) to no-panic fuzz targets
  to exercise the missing-content-type code path
- Relax chunk-split assertion to tolerate whitespace-only drift
