+++ components/rust-converter/fuzz/fuzz_targets/fuzz_streaming_no_panic.rs
@@ -5,8 +5,16 @@ use nginx_markdown_converter::streaming::budget::MemoryBudget;
 use nginx_markdown_converter::streaming::converter::StreamingConverter;
 
 fuzz_target!(|data: &[u8]| {
+    let (set_content_type, payload) = match data.split_first() {
+        Some((flag, rest)) => ((flag & 1) == 0, rest),
+        None => (true, data),
+    };
+
     let mut conv = StreamingConverter::new(ConversionOptions::default(), MemoryBudget::default());
-    conv.set_content_type(Some("text/html; charset=UTF-8".to_string()));
-    let _ = conv.feed_chunk(data);
+    if set_content_type {
+        conv.set_content_type(Some("text/html; charset=UTF-8".to_string()));
+    }
+
+    let _ = conv.feed_chunk(payload);
     let _ = conv.finalize();
 });
